Move `run_debug_ipc_service` to `service` module
